Small bugfixes, assertion correction and improvements for the isolate environment.
git-svn-id: https://llvm.org/svn/llvm-project/vmkit/branches/isolate@55595 91177308-0d34-0410-b5e6-96231b3b80d8

+++ lib/JnJVM/Isolate/IsolateCommonClass.h
@@ -188,7 +188,7 @@ public:
                                   UserCommonClass*& fieldCl);
   
   
-  uint64 getVirtualSize() {
+  uint32 getVirtualSize() {
     return virtualSize;
   }
 
@@ -319,7 +319,7 @@ public:
     return ctpInfo;
   }
 
-  uint64 getStaticSize() {
+  uint32 getStaticSize() {
     return ((Class*)classDef)->getStaticSize();
   }
